+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * @dev (????-??-??)
   * Logisim has now an internal font-chooser to comply to the font-values used
+  * You can now swap the placement of main canvas and component tree/properties pane
 
 * v3.7.0 (2021-10-12)
   * Reworked the slider component in the I/O extra library